+++ src/hdf/object/fits/FitsDatatype.java
@@ -29,7 +29,7 @@ import nom.tam.fits.BasicHDU;
 public class FitsDatatype extends Datatype
 {
     private static final long serialVersionUID = 6545936196104493765L;
-    private int nativeType;
+    private long nativeType;
 
     /**
      * Create an Datatype with specified class, size, byte order and sign.
@@ -44,7 +44,7 @@ public class FitsDatatype extends Datatype
      * <LI>to create 64-bit double<br>
      * FitsDatatype type = new H5Dataype(CLASS_FLOAT, 8, NATIVE, -1);
      * </OL>
-     * <p>
+     *
      * @param tclass the class of the datatype.
      * @param tsize the size of the datatype in bytes.
      * @param torder the order of the datatype.
@@ -56,10 +56,10 @@ public class FitsDatatype extends Datatype
 
     /**
      * Create a Datatype with a given fits native datatype.
-     * <p>
+     *
      * @param theType the fits native datatype.
      */
-    public FitsDatatype(int theType) {
+    public FitsDatatype(long theType) {
         super(-1);
         nativeType = theType;
         fromNative(0);
@@ -75,12 +75,12 @@ public class FitsDatatype extends Datatype
      * Allocate an one-dimensional array of byte, short, int, long, float, double,
      * or String to store data retrieved from an fits file based on the given
      * fits datatype and dimension sizes.
-     * <p>
+     *
      * @param dtype the fits datatype.
      * @param size the total size of the array.
      * @return the array object if successful and null otherwise.
      */
-    public static Object allocateArray(int dtype, int size) throws OutOfMemoryError
+    public static Object allocateArray(long dtype, int size) throws OutOfMemoryError
     {
         Object data = null;
 
@@ -88,7 +88,7 @@ public class FitsDatatype extends Datatype
             return null;
         }
 
-        switch (dtype) {
+        switch ((int)dtype) {
             case BasicHDU.BITPIX_BYTE:
                 data = new byte[size];
                 break;
@@ -121,13 +121,13 @@ public class FitsDatatype extends Datatype
 
     /**
      * Translate fits datatype identifier into FitsDatatype.
-     * <p>
+     *
      * @param nativeID the fits native datatype.
      */
     @Override
-    public void fromNative(int dtype)
+    public void fromNative(long dtype)
     {
-        switch (dtype) {
+        switch ((int)dtype) {
             case BasicHDU.BITPIX_BYTE:
                 datatypeClass = CLASS_INTEGER;
                 datatypeSize = 1;
@@ -160,7 +160,7 @@ public class FitsDatatype extends Datatype
     public String getDatatypeDescription() {
         String description = "Unknown data type.";
 
-        switch (nativeType) {
+        switch ((int)nativeType) {
             case BasicHDU.BITPIX_BYTE:
                 description = "8-bit integer";
                 break;
@@ -182,13 +182,13 @@ public class FitsDatatype extends Datatype
             default:
                 if (datatypeClass==Datatype.CLASS_STRING) {
                     description = "String";
-                } 
+                }
                 else if (datatypeClass==Datatype.CLASS_CHAR) {
                     description = "Char";
-                } 
+                }
                 else if (datatypeClass==Datatype.CLASS_INTEGER) {
                     description = "Integer";
-                } 
+                }
                 else if (datatypeClass==Datatype.CLASS_FLOAT) {
                     description = "Float";
                 }
@@ -206,25 +206,25 @@ public class FitsDatatype extends Datatype
 
     // implementing Datatype
     @Override
-    public int toNative() {
+    public long toNative() {
         if (datatypeClass == CLASS_INTEGER) {
             if (datatypeSize == 1) {
                 nativeType = BasicHDU.BITPIX_BYTE;
-            } 
+            }
             else if (datatypeSize == 2) {
                 nativeType = BasicHDU.BITPIX_SHORT;
-            } 
+            }
             else if (datatypeSize == 4) {
                 nativeType = BasicHDU.BITPIX_INT;
-            } 
+            }
             else if (datatypeSize == 8) {
                 nativeType = BasicHDU.BITPIX_LONG;
             }
-        } 
+        }
         else if (datatypeClass == CLASS_FLOAT) {
             if (datatypeSize == 4) {
                 nativeType = BasicHDU.BITPIX_FLOAT;
-            } 
+            }
             else if (datatypeSize == 8) {
                 nativeType = BasicHDU.BITPIX_DOUBLE;
             }
@@ -238,10 +238,10 @@ public class FitsDatatype extends Datatype
      * @see hdf.object.Datatype#close(int)
      */
     @Override
-    public void close(int id) {;}
+    public void close(long id) {;}
 
   //Implementing DataFormat
     public List getMetadata(int... attrPropList) throws Exception {
         throw new UnsupportedOperationException("getMetadata(int... attrPropList) is not supported");
-    }    
+    }
 }
